refactor(UserInInvitation): move validation to UserInInvitation file
